refactor(sdk): centralize Hinnant civil-date math in util::time
The days-from-civil / civil-from-days primitives were copied across
util::time (inlined in parse_iso_ms), query_verbs::mod, and reader::opencode.
Promote ymd_to_days + days_to_ymd to pub(crate) in util::time and route the
callers there:

- parse_iso_ms now calls ymd_to_days instead of inlining it.
- query_verbs::ymd_to_days keeps only its out-of-range guard, deferring the
  math to the shared primitive; its days_to_ymd copy is replaced by the import.
- reader::opencode::ms_to_iso uses the shared days_to_ymd (the {:04} year
  format works on the i64 year identically to the old i32).

Identical math, behavior-preserving; build warning-free, clippy clean, 990
tests pass.
