+++ crates/relayburn-sdk/src/query_verbs/mod.rs
@@ -259,36 +259,17 @@ fn format_iso_z_ms(secs: i64, millis: u32) -> String {
     format!("{year:04}-{month:02}-{day:02}T{hour:02}:{minute:02}:{second:02}.{millis:03}Z")
 }
 
-/// Civil-date → days-from-Unix-epoch (Howard Hinnant's algorithm, proleptic
-/// Gregorian). Inverse of [`days_to_ymd`].
+/// Range-checking wrapper over [`crate::util::time::ymd_to_days`]: rejects
+/// out-of-range month/day (since this parses untrusted `since` strings),
+/// then defers to the shared Hinnant primitive.
 fn ymd_to_days(year: i64, month: u32, day: u32) -> Option<i64> {
     if !(1..=12).contains(&month) || !(1..=31).contains(&day) {
         return None;
     }
-    let m = month as i64;
-    let d = day as i64;
-    let y = if m <= 2 { year - 1 } else { year };
-    let era = if y >= 0 { y } else { y - 399 } / 400;
-    let yoe = (y - era * 400) as u64;
-    let mp = if m > 2 { m - 3 } else { m + 9 } as u64;
-    let doy = (153 * mp + 2) / 5 + (d as u64) - 1;
-    let doe = yoe * 365 + yoe / 4 - yoe / 100 + doy;
-    Some(era * 146_097 + (doe as i64) - 719_468)
+    Some(crate::util::time::ymd_to_days(year, month, day))
 }
 
-fn days_to_ymd(days_from_epoch: i64) -> (i64, u32, u32) {
-    let z = days_from_epoch + 719_468;
-    let era = if z >= 0 { z } else { z - 146_096 } / 146_097;
-    let doe = (z - era * 146_097) as u64;
-    let yoe = (doe - doe / 1_460 + doe / 36_524 - doe / 146_096) / 365;
-    let y = yoe as i64 + era * 400;
-    let doy = doe - (365 * yoe + yoe / 4 - yoe / 100);
-    let mp = (5 * doy + 2) / 153;
-    let d = doy - (153 * mp + 2) / 5 + 1;
-    let m = if mp < 10 { mp + 3 } else { mp - 9 };
-    let year = if m <= 2 { y + 1 } else { y };
-    (year, m as u32, d as u32)
-}
+use crate::util::time::days_to_ymd;
 
 // ---------------------------------------------------------------------------
 // time-bucketing — shared by `--bucket` on summary / compare / hotspots / overhead
