+++ crates/relayburn-sdk/src/util/time.rs
@@ -50,6 +50,17 @@ pub(crate) fn parse_iso_ms(s: &str) -> Option<i64> {
         }
         millis = frac.parse().ok()?;
     }
+    let days_from_epoch = ymd_to_days(year, month, day);
+    let secs =
+        days_from_epoch * 86_400 + (hour as i64) * 3_600 + (minute as i64) * 60 + (second as i64);
+    Some(secs * 1_000 + millis)
+}
+
+/// Civil date → days from the Unix epoch (Howard Hinnant's proleptic-Gregorian
+/// `days_from_civil`). Inverse of [`days_to_ymd`]. Does **not** range-check its
+/// inputs — callers that accept untrusted month/day values guard the range
+/// themselves before calling.
+pub(crate) fn ymd_to_days(year: i64, month: u32, day: u32) -> i64 {
     let m = month as i64;
     let d = day as i64;
     let y = if m <= 2 { year - 1 } else { year };
@@ -58,10 +69,23 @@ pub(crate) fn parse_iso_ms(s: &str) -> Option<i64> {
     let mp = if m > 2 { m - 3 } else { m + 9 } as u64;
     let doy = (153 * mp + 2) / 5 + (d as u64) - 1;
     let doe = yoe * 365 + yoe / 4 - yoe / 100 + doy;
-    let days_from_epoch = era * 146_097 + (doe as i64) - 719_468;
-    let secs =
-        days_from_epoch * 86_400 + (hour as i64) * 3_600 + (minute as i64) * 60 + (second as i64);
-    Some(secs * 1_000 + millis)
+    era * 146_097 + (doe as i64) - 719_468
+}
+
+/// Days from the Unix epoch → `(year, month, day)` (Howard Hinnant's
+/// `civil_from_days`). Inverse of [`ymd_to_days`].
+pub(crate) fn days_to_ymd(days_from_epoch: i64) -> (i64, u32, u32) {
+    let z = days_from_epoch + 719_468;
+    let era = if z >= 0 { z } else { z - 146_096 } / 146_097;
+    let doe = (z - era * 146_097) as u64;
+    let yoe = (doe - doe / 1_460 + doe / 36_524 - doe / 146_096) / 365;
+    let y = yoe as i64 + era * 400;
+    let doy = doe - (365 * yoe + yoe / 4 - yoe / 100);
+    let mp = (5 * doy + 2) / 153;
+    let d = doy - (153 * mp + 2) / 5 + 1;
+    let m = if mp < 10 { mp + 3 } else { mp - 9 };
+    let year = if m <= 2 { y + 1 } else { y };
+    (year, m as u32, d as u32)
 }
 
 #[cfg(test)]
